Rename to RemoveMixedDocblockOverruledByNativeTypeRector
