+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveMixedDocblockOverruledByNativeTypeRector/Fixture/remove_param_and_return_mixed.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockOverruledByNativeTypeRector\Fixture;
 
 final class RemoveParamAndReturnMixed
 {
@@ -18,7 +18,7 @@ final class RemoveParamAndReturnMixed
 -----
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockOverruledByNativeTypeRector\Fixture;
 
 final class RemoveParamAndReturnMixed
 {
