+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveMixedDocblockOverruledByNativeTypeRector/Fixture/skip_return_no_native_type.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockOverruledByNativeTypeRector\Fixture;
 
 final class SkipReturnNoNativeType
 {
