+++ rules-tests/DeadCode/Rector/ClassMethod/RemoveMixedDocblockOverruledByNativeTypeRector/RemoveMixedDocblockOverruledByNativeTypeRectorTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockRector;
+namespace Rector\Tests\DeadCode\Rector\ClassMethod\RemoveMixedDocblockOverruledByNativeTypeRector;
 
 use Iterator;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-final class RemoveMixedDocblockRectorTest extends AbstractRectorTestCase
+final class RemoveMixedDocblockOverruledByNativeTypeRectorTest extends AbstractRectorTestCase
 {
     #[DataProvider('provideData')]
     public function test(string $filePath): void
