--- rules-tests/DeadCode/Rector/ClassMethod/RemoveMixedDocblockRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\DeadCode\Rector\ClassMethod\RemoveMixedDocblockRector;
-
-return RectorConfig::configure()
-    ->withRules([RemoveMixedDocblockRector::class]);
